requirements.txt
awsebcli==3.5.2
blessed==1.9.5
botocore==1.1.10
cement==2.4.0
colorama==0.3.3
Django==1.8.4
django-sticky-uploads==0.4.0
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
gunicorn==19.3.0
image==1.4.1
jmespath==0.7.1
pathspec==0.3.3
Pillow==2.9.0
psycopg2==2.6.1
python-dateutil==2.4.2
PyYAML==3.11
requests==2.6.2
six==1.9.0
texttable==0.8.3
wcwidth==0.1.4
websocket-client==0.32.0
wheel==0.24.0
